Move out Exposer class
Fixes swagger-php warning:

  $ ./vendor/bin/openapi --version 3.1.0 src/Controller/VHosts/Api/ src/Model/Behavior/ExposedOnApiBehavior.php -o webroot/api/openapi.json
  Warning: Skipping unknown App\Model\Behavior\Exposer
